依存関係の修正
