+++ app/javascript/controllers/simulation_mode_controller.js
@@ -25,9 +25,9 @@ export default class extends Controller {
   close(event) {
     if (event.detail.success) {
       this.closeDetail();
-      const url = event.detail.fetchResponse.response.url;
-      const redirectUrl = url.replace(/\/\d+$/, "");
-      window.Turbo.visit(redirectUrl);
+      // const url = event.detail.fetchResponse.response.url;
+      // const redirectUrl = url.replace(/\/\d+$/, "");
+      // window.Turbo.visit(redirectUrl);
     }
   }
